requirements.txt
requests==2.31.0
pyyaml==6.0.1
pyshark==0.6
scapy==2.5.0
jinja2==3.1.3
weasyprint==62.3
pytest==8.1.1
pydantic==2.12.5
colorlog==6.8.2
tabulate==0.9.0